refactor(filters): move str comparisions to a seperate file in utils
