+++ pkg/util/util.go
@@ -112,40 +112,6 @@ func Hash(s string) string {
 	return hex.EncodeToString(hasher.Sum(nil))
 }
 
-// Contains is a util function which returns true if one key is present in array
-// else it returns false
-func Contains(s []string, k string) bool {
-	for _, e := range s {
-		if e == k {
-			return true
-		}
-	}
-	return false
-}
-
-// ContainsIgnoredCase is a util function which returns true if one key is present
-// in array else it returns false. This function is not case sensitive.
-func ContainsIgnoredCase(s []string, k string) bool {
-	for _, e := range s {
-		if strings.ToLower(e) == strings.ToLower(k) {
-			return true
-		}
-	}
-	return false
-}
-
-// MatchIgnoredCase is a util function which returns true if any of the keys
-// are present as a string in given string - s
-// This function is not case sensitive.
-func MatchIgnoredCase(keys []string, s string) bool {
-	for _, k := range keys {
-		if strings.Contains(strings.ToLower(s), strings.ToLower(k)) {
-			return true
-		}
-	}
-	return false
-}
-
 // StateStatus returns enable if state is true and disable if state is false
 func StateStatus(state bool) string {
 	var status = map[bool]string{
