+++ pkg/util/util_test.go
@@ -134,61 +134,6 @@ func TestCheckErr(t *testing.T) {
 	CheckErr(nil, handlerFunc)
 }
 
-func TestContains(t *testing.T) {
-	diskList := make([]string, 0)
-	diskList = append(diskList, "Key1")
-	diskList = append(diskList, "Key3")
-	tests := map[string]struct {
-		diskName string
-		expected bool
-	}{
-		"giving a key which is not present in slice": {diskName: "Key0", expected: false},
-		"giving a key which is present in slice":     {diskName: "Key3", expected: true},
-	}
-	for name, test := range tests {
-		t.Run(name, func(t *testing.T) {
-			assert.Equal(t, test.expected, Contains(diskList, test.diskName))
-		})
-	}
-}
-
-func TestContainsIgnoredCase(t *testing.T) {
-	diskList := make([]string, 0)
-	diskList = append(diskList, "Key1")
-	diskList = append(diskList, "Key3")
-	tests := map[string]struct {
-		diskName string
-		expected bool
-	}{
-		"giving a key which is not present in slice": {diskName: "keY0", expected: false},
-		"giving a key which is present in slice":     {diskName: "KEy3", expected: true},
-	}
-	for name, test := range tests {
-		t.Run(name, func(t *testing.T) {
-			assert.Equal(t, test.expected, ContainsIgnoredCase(diskList, test.diskName))
-		})
-	}
-}
-
-func TestMatchIgnoredCase(t *testing.T) {
-	mkList := make([]string, 0)
-	mkList = append(mkList, "loop")
-	mkList = append(mkList, "/dev/sr0")
-	tests := map[string]struct {
-		diskPath string
-		expected bool
-	}{
-		"diskPath contains one of the keys ": {diskPath: "/dev/loop0", expected: true},
-		"diskPath matches complete key":      {diskPath: "/dev/sr0", expected: true},
-		"diskPath does not match any keys":   {diskPath: "/dev/sdb", expected: false},
-	}
-	for name, test := range tests {
-		t.Run(name, func(t *testing.T) {
-			assert.Equal(t, test.expected, MatchIgnoredCase(mkList, test.diskPath))
-		})
-	}
-}
-
 func TestStateStatus(t *testing.T) {
 	tests := map[string]struct {
 		status string
